refactor: Vulnerability components moved to subfolder
