+++ src/lib/singleview/vulnerabilities/vulnerability/Acknowledgements.svelte
