+++ src/lib/singleview/vulnerabilities/vulnerability/Flags.svelte
