+++ src/lib/singleview/vulnerabilities/vulnerability/GeneralSection.svelte
