+++ src/lib/singleview/vulnerabilities/vulnerability/ID.svelte
