+++ src/lib/singleview/vulnerabilities/vulnerability/Vulnerability.svelte
